# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Open-source drug information gateway. Provides NDC lookup, therapeutic class search, drug interactions, RxNorm fuzzy search, and structured product label data — all through a clean REST API with Redis caching, per-key rate limiting, and circuit breaker protection.\n\n## Getting Started\n1. **Create an API key:** `POST /admin/keys` with `Authorization: Bearer {ADMIN_SECRET}` and body `{\"app_name\": \"my-app\", \"origins\": [\"*\"], \"rate_limit\": 1000}`\n2. **Search for a drug:** `GET /v1/drugs/autocomplete?q=lipit` with header `X-API-Key: {your_key}` — returns matching drug names\n3. **Look up details:** `GET /v1/drugs/class?name=atorvastatin` — returns therapeutic classes and brand names\n4. **Get interactions:** `GET /v1/drugs/info?name=warfarin` — returns FDA label interaction warnings\n5. **Check multi-drug interactions:** `POST /v1/drugs/interactions` with `{\"drugs\": [{\"name\": \"warfarin\"}, {\"name\": \"aspirin\"}]}`",
        "title": "drug-gate API",
        "contact": {
            "name": "drug-gate on GitHub",
            "url": "https://github.com/finish06/drug-gate"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/finish06/drug-gate/blob/main/LICENSE"
        },
        "version": "0.9.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/drugs/rxnorm/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Resolves a drug name via RxNorm approximate match, then assembles NDCs, generic equivalents, and all related concepts into a single unified response. This is a convenience endpoint that combines the results of search, NDCs, generics, and related into one call. Use this when you need a complete drug profile from a single request.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rxnorm"
                ],
                "summary": "Get unified drug profile",
                "parameters": [
                    {
                        "type": "string",
                        "example": "metformin",
                        "description": "Drug name (generic or brand)",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormProfile"
                        }
                    },
                    "400": {
                        "description": "Missing name parameter",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Drug not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/rxnorm/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Performs an approximate-match search via the RxNorm API and returns up to 5 candidates ranked by score, each with an RxCUI. When no exact matches are found, spelling suggestions are included to help correct the query. Use this as the entry point for RxNorm workflows before calling NDC, generic, or related endpoints.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rxnorm"
                ],
                "summary": "Search drugs by name (RxNorm)",
                "parameters": [
                    {
                        "type": "string",
                        "example": "lisinopril",
                        "description": "Drug name to search for",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormSearchResult"
                        }
                    },
                    "400": {
                        "description": "Missing name parameter",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "No drugs found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/rxnorm/{rxcui}/generics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns generic product information for the given RxNorm concept identifier, including ingredient names and dose form groups. Use this endpoint to find generic equivalents of a branded drug after resolving its RxCUI via the search endpoint.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rxnorm"
                ],
                "summary": "Get generic products for an RxCUI",
                "parameters": [
                    {
                        "type": "string",
                        "example": "314076",
                        "description": "RxNorm concept unique identifier",
                        "name": "rxcui",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormGenericResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "RxCUI not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/rxnorm/{rxcui}/ndcs": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all National Drug Code (NDC) identifiers associated with the given RxNorm concept. Use this to map an RxCUI obtained from the search endpoint to specific packaged products. Results are cached with a long TTL since NDC-to-RxCUI mappings change infrequently.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rxnorm"
                ],
                "summary": "Get NDCs for an RxCUI",
                "parameters": [
                    {
                        "type": "string",
                        "example": "314076",
                        "description": "RxNorm concept unique identifier",
                        "name": "rxcui",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormNDCResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "RxCUI not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/drugs/rxnorm/{rxcui}/related": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all related RxNorm concepts grouped by relationship type: ingredients, brand names, dose forms, clinical drugs, and branded drugs. Returns 404 only when all groups are empty, indicating an unknown RxCUI. Individual empty groups within a valid response are normal.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rxnorm"
                ],
                "summary": "Get related concepts for an RxCUI",
                "parameters": [
                    {
                        "type": "string",
                        "example": "314076",
                        "description": "RxNorm concept unique identifier",
                        "name": "rxcui",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormRelatedResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid API key",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "RxCUI not found or has no related concepts",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Upstream service unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_finish06_drug-gate_internal_model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "not_found"
                },
                "message": {
                    "type": "string",
                    "example": "No drug found for NDC 99999-9999"
                }
            }
        },
...
        "github_com_finish06_drug-gate_internal_model.RxNormCandidate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "atorvastatin"
                },
                "rxcui": {
                    "type": "string",
                    "example": "153165"
                },
                "score": {
                    "type": "integer",
                    "example": 94
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.RxNormConcept": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "atorvastatin calcium 10 MG Oral Tablet"
                },
                "rxcui": {
                    "type": "string",
                    "example": "314076"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.RxNormGenericResponse": {
            "type": "object",
            "properties": {
                "generics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                    }
                },
                "rxcui": {
                    "type": "string",
                    "example": "314076"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.RxNormNDCResponse": {
            "type": "object",
            "properties": {
                "ndcs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "00069-3150-30"
                    ]
                },
                "rxcui": {
                    "type": "string",
                    "example": "314076"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.RxNormProfile": {
            "type": "object",
            "properties": {
                "brand_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Glucophage"
                    ]
                },
                "generic": {
                    "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                },
                "name": {
                    "type": "string",
                    "example": "metformin"
                },
                "ndcs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "00087-6060-05"
                    ]
                },
                "query": {
                    "type": "string",
                    "example": "metformin"
                },
                "related": {
                    "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormRelatedResponse"
                },
                "rxcui": {
                    "type": "string",
                    "example": "6809"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.RxNormRelatedResponse": {
            "type": "object",
            "properties": {
                "brand_names": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                    }
                },
                "branded_drugs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                    }
                },
                "clinical_drugs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                    }
                },
                "dose_forms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                    }
                },
                "ingredients": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormConcept"
                    }
                },
                "rxcui": {
                    "type": "string",
                    "example": "314076"
                }
            }
        },
        "github_com_finish06_drug-gate_internal_model.RxNormSearchResult": {
            "type": "object",
            "properties": {
                "candidates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_finish06_drug-gate_internal_model.RxNormCandidate"
                    }
                },
                "query": {
                    "type": "string",
                    "example": "lipitor"
                },
                "suggestions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "lipitor"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AdminAuth": {
            "description": "Admin bearer token set via `ADMIN_SECRET` environment variable. Format: `Bearer {secret}`. Required for all `/admin/*` endpoints.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "ApiKeyAuth": {
            "description": "Publishable API key for frontend applications. Create one via `POST /admin/keys`. Include in all `/v1/*` requests.",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    },
...
}